Fixed C++20 compatibility issues. Reduced warnings.

+++ src/client/render/Window.cpp
@@ -22,7 +22,7 @@ Window::Window(const char* title)
 	m_window = std::make_unique<sf::RenderWindow>(vm, title);
 	m_window->setFramerateLimit(60);
 
-	sf::FloatRect visibleArea(0.f, 0.f, width, height);
+	sf::FloatRect visibleArea(0.f, 0.f, static_cast<float>(width), static_cast<float>(height));
 	m_window->setView(sf::View(visibleArea));
 }
